Check Stan model for required files and syntax errors

Ensure that a bbi_stan_model object has all the files necessary to submit the model (see ?bbr_stan for details). By default, also check the syntax of that the .stan file.

check_stan_model(.mod, .syntax = TRUE, .error = FALSE)

Arguments

.mod

A bbi_stan_model object

.syntax

Check the syntax of the model by calling CmdStanModel$check_syntax().

.error

If FALSE, the default, display a message for any problems found. If TRUE, signal an error.

Value

Invisibly return FALSE if any problems are found (only relevant when .error is FALSE) and TRUE otherwise.

Details

Will look for the following:

build_path_from_model(.mod, "-standata.R")

build_path_from_model(.mod, "-stanargs.R")

build_path_from_model(.mod, ".stan")

build_path_from_model(.mod, "-init.R"), unless it is a generated quantities model (i.e. a bbi_stan_gq_model subclass).

build_path_from_model(.mod, "-fitted-params.R"), if it is a generated quantities model (i.e. a bbi_stan_gq_model subclass).