Check that Stan model is up to date

The details of which files are "model files" and which are "data files" are detailed in the bbr::check_up_to_date() docs.

# S3 method for bbi_stan_model
check_up_to_date(.bbi_object, .build_data = TRUE, ...)

# S3 method for bbi_stan_gq_model
check_up_to_date(.bbi_object, .build_data = TRUE, ...)

# S3 method for bbi_stan_summary
check_up_to_date(.bbi_object, .build_data = TRUE, ...)

# S3 method for bbi_stan_gq_summary
check_up_to_date(.bbi_object, .build_data = TRUE, ...)

Arguments

.bbi_object

the object to check. Could be a bbi_{.model_type}_model object, a bbi_{.model_type}_summary object, or a bbi_log_df tibble.

.build_data

If TRUE, the default, run -standata.R and save the output to a temp file and check the hash of the temp file against the bbi_config.json hash. This option actually runs the code and, importantly, verifies that the input data to -standata.R has not changed either. If FALSE, check the hashes of the -standata.R and -standata.json against the hashes in bbi_config.json but do not run the -standata.R script. This option is less secure and primarily exists for quicker checking if building the data is time consuming for certain models.

...

Arguments passed through (currently none).

Details

The model files:

<run>.stan

<run>-stanargs.R

<run>-init.R, if the model is not a standalone generated quantities model

<run>-fitted-params.R, if the model is a standalone generated quantities model (i.e. a model that inherits from bbi_stan_gq_model).

The data files (see .build_data argument):

<run>-standata.R

<run>-standata.json

For standalone generated quantities, the bbi_config.json files in the output directory of the models listed in the gq_parent fields are also considered "data" files. If a gq_parent is re-submitted and anything recorded in bbi_config.json changes, the linked "stan_gq" model is considered out of date.