Summarize a bbi_stan_model

There is currently no bbr::model_summary() method for summarizing Stan outputs. Call read_fit_model() to return the resulting fit object, which has various methods for summarizing the outputs. See the ?cmdstanr::CmdStanMCMC docs for methods and information on this object.

# S3 method for bbi_stan_model
model_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)

Arguments

.mod

Model to summarize.

.bbi_args

A named list specifying arguments to pass to bbi formatted like list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4). See print_bbi_args() for full list of options.

...

args passed through to bbi_exec()

.dry_run

show what the command would be without actually running it