Check if a NONMEM Bayes model has finished running

This function implements the bbr::check_nonmem_finished() method for NONMEM Bayes models. To decide if a model is finished, it runs bbr::check_nonmem_finished() on each of the chain submodels.

# S3 method for class 'bbi_nmbayes_model'
check_nonmem_finished(.mod, ...)

Arguments

.mod

A bbi_nmbayes_model object.

...

Arguments passed to method call on each chain submodel.

Value

TRUE if all of the chain models have finished, and FALSE otherwise.

Details

As with regular NONMEM models, the model is not considered finished if the output directory does not exist.