Inspect first and last lines of select files from a NONMEM Bayes model

tail_lst() and tail_output() provide bbr::tail_lst() and bbr::tail_output() methods for bbi_nmbayes_model objects. Compared to the methods for bbi_nonmem_model objects, the key difference is that these operate on the *.lst or OUTPUT file of each chain submodel.

# S3 method for bbi_nmbayes_model
tail_lst(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)

# S3 method for bbi_nmbayes_model
tail_output(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)

Arguments

.mod

Model object or path associated with file.

.head

Integer for number of lines to read from the top of the file

.tail

Integer for number of lines to read from the bottom of the file

.print

If TRUE, the default, print resulting head and tail to console.

.return

If TRUE, return the head and tail results for each chain. See ‘Value’ section for more details.

...

Arguments passed to method call on each chain submodel.

Value

If .return is TRUE, return a list of character vectors, with an element for each chain. The character vector is comprised of the first and last lines of the file under a chain subdirectory. For tail_output(), the element will be NULL if 1) the model has started to run but hasn't yet created the OUTPUT file or 2) the model has finished running.

See also